feat: add `logRequestDetails` to middleware
